requirements.txt
django=5.2
djangorestframework=3.16.0
django-cors-headers=4.7.0
pillow=11.2.1
psycopg2-binary=2.9.11
requests=2.32.3